95 material

+++ app/src/main/kotlin/ru/erdenian/studentassistant/ui/lessonseditor/LessonsEditorActivity.kt
@@ -13,15 +13,16 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.viewModelScope
 import androidx.viewpager.widget.PagerTabStrip
 import androidx.viewpager.widget.ViewPager
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import kotlinx.coroutines.launch
-import org.jetbrains.anko.alert
 import org.jetbrains.anko.startActivity
 import ru.erdenian.studentassistant.R
 import ru.erdenian.studentassistant.entity.Semester
 import ru.erdenian.studentassistant.ui.lessoneditor.LessonEditorActivity
 import ru.erdenian.studentassistant.ui.semestereditor.SemesterEditorActivity
 import ru.erdenian.studentassistant.utils.getColorCompat
+import ru.erdenian.studentassistant.utils.getResId
 import ru.erdenian.studentassistant.utils.requireViewByIdCompat
 import ru.erdenian.studentassistant.utils.setColor
 
@@ -70,8 +71,8 @@ class LessonsEditorActivity : AppCompatActivity(R.layout.activity_lessons_editor
             adapter = LessonsEditorPagerAdapter(supportFragmentManager)
         }
         requireViewByIdCompat<PagerTabStrip>(R.id.alse_by_weekdays_pager_tab_strip).apply {
-            setTextColor(getColorCompat(R.color.primary))
-            setTabIndicatorColorResource(R.color.primary)
+            setTextColor(getColorCompat(getResId(R.attr.colorPrimary)))
+            setTabIndicatorColorResource(getResId(R.attr.colorPrimary))
         }
 
         // TODO: 13.11.2016 добавить заполнение списка пар по датам
@@ -107,12 +108,13 @@ class LessonsEditorActivity : AppCompatActivity(R.layout.activity_lessons_editor
             true
         }
         R.id.mlse_delete_semester -> {
-            alert(R.string.lsea_delete_message) {
-                positiveButton(R.string.lsea_delete_yes) {
+            MaterialAlertDialogBuilder(this)
+                .setMessage(R.string.lsea_delete_message)
+                .setPositiveButton(R.string.lsea_delete_yes) { _, _ ->
                     viewModel.viewModelScope.launch { viewModel.deleteSemester() }
                 }
-                negativeButton(R.string.lsea_delete_no) {}
-            }.show()
+                .setNegativeButton(R.string.lsea_delete_no, null)
+                .show()
             true
         }
         else -> false
